+++ crates/test-macro/src/lib.rs
@@ -74,10 +74,11 @@ pub fn wasm_bindgen_test(
     // later slurp up all of these functions and pass them as arguments to the
     // main test harness. This is the entry point for all tests.
     let name = format_ident!("__wbgt_{}_{}", ident, CNT.fetch_add(1, Ordering::SeqCst));
+    let wasm_bindgen_path = attributes.wasm_bindgen_path;
     tokens.extend(
         quote! {
             #[no_mangle]
-            pub extern "C" fn #name(cx: &::wasm_bindgen_test::__rt::Context) {
+            pub extern "C" fn #name(cx: &#wasm_bindgen_path::__rt::Context) {
                 let test_name = ::core::concat!(::core::module_path!(), "::", ::core::stringify!(#ident));
                 #test_body
             }
@@ -187,15 +188,26 @@ fn compile_error(span: Span, msg: &str) -> proc_macro::TokenStream {
     quote_spanned! { span => compile_error!(#msg); }.into()
 }
 
-#[derive(Default)]
 struct Attributes {
     r#async: bool,
+    wasm_bindgen_path: syn::Path,
+}
+
+impl Default for Attributes {
+    fn default() -> Self {
+        Self {
+            r#async: false,
+            wasm_bindgen_path: syn::parse_quote!(::wasm_bindgen_test),
+        }
+    }
 }
 
 impl Attributes {
     fn parse(&mut self, meta: syn::meta::ParseNestedMeta) -> syn::parse::Result<()> {
         if meta.path.is_ident("async") {
             self.r#async = true;
+        } else if meta.path.is_ident("crate") {
+            self.wasm_bindgen_path = meta.value()?.parse::<syn::Path>()?;
         } else {
             return Err(meta.error("unknown attribute"));
         }